pyproject.toml
[project]
name = "scheduler-service"
version = "0.3.0"
description = ""
authors = [
    {name = "CoCongV", email = "cong.lv.yx@gmail.com"},
    {name = "Fuxin Hao", email = "haofxpro@gmail.com"},
]
dependencies = [
    "uvicorn~=0.38.0",
    "click~=8.3.1",
    "asyncpg~=0.23",
    "tortoise-orm (>=0.21.0)",
    "passlib~=1.7",
    "pyjwt~=2.10.1",
    "dramatiq[redis, watch]~=2.0.0",
    "httpx~=0.25.0",
    "fastapi (>=0.121.3,<0.122.0)",
    "pydantic[email] (>=2.12.4,<3.0.0)",
    "apscheduler (>=3.11.1,<4.0.0)",
    "redis (>=5.0.0,<6.0.0)",
    "dramatiq-abort (>=1.2.1,<2.0.0)",
    "aerich (>=0.9.2,<0.10.0)",

]
requires-python = ">=3.14,<4.0"
dynamic = ["classifiers"]
license = {text = "MIT"}

[project.urls]
homepage = "https://github.com/CoCongV/scheduler-service#"


[project.scripts]
scheduler = "scheduler_service.cli:scheduler"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{include = "scheduler_service"}]

[tool.coverage.run]
omit = ["scheduler_service/cli.py"]

[dependency-groups]
dev = [
    "pylint (>=4.0.3,<5.0.0)",
    "ipython (>=9.7.0,<10.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "httpx (>=0.25.0,<1.0.0)",
    "pytest (>=9.0.1,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "pytest-mock (>=3.12.0,<4.0.0)",
    "autopep8 (>=2.3.2,<3.0.0)",
]

[tool.aerich]
tortoise_orm = "scheduler_service.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
